requirements.txt
Django==3.1
psycopg2-binary==2.8.5
Pillow==8.0.1
djangorestframework==3.11
python-decouple==3